requirements.txt
python-telegram-bot==21.6
requests
python-dotenv
feedparser
anthropic
apscheduler
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytz
